Refactored logger color scheme and removed testing logs
Refactored the logging utility by replacing the colour RGB system with a hexadecimal system by introducing a colours constant. This will make it easier to maintain and understand the logger utility in the future. Also removed redundant testing log generation and some unnecessary comments for cleaner and more efficient code.